pyproject.toml
[tool.poetry]
name = "test-assignment"
version = "0.1.0"
description = ""
authors = ["Oleksandr Kozlov <kozlov2777@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^8.3.5"
selenium = "^4.31.0"
faker = "^37.1.0"
pydantic = "^2.11.3"
httpx = "^0.28.1"
webdriver-manager = "^4.0.2"
allure-pytest = "^2.14.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.2.0"
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
